Split step5 code
