+++ rootkit/Kbuild
@@ -1,3 +1,3 @@
 obj-m := yarin_module.o
 ccflags-y += -I$(src)/include
-yarin_module-y := yarin_module_main.o helpers.o step2.o step3.o
+yarin_module-y := yarin_module_main.o helpers.o step2.o step3.o step5.o
